requirements.txt
PySide6-Essentials==6.6.3

